Revert "Rename some files"
This reverts commit acb076873a0538de9b22d657f6398930f9f0ada4.
